chore: move common helpers into canonical package
