+++ easyuse_anima/common/__init__.py
@@ -1,0 +1,3 @@
+"""Cross-feature primitive package boundary."""
+
+__all__ = ()
