+++ easyuse_anima/common/values.py
@@ -1,0 +1,51 @@
+"""Domain-independent value coercion helpers."""
+
+
+def _single_value(value):
+    if isinstance(value, (list, tuple)):
+        if not value:
+            return None
+        return value[0]
+    return value
+
+
+def _as_bool(value, default: bool = False) -> bool:
+    value = _single_value(value)
+    if value is None:
+        return default
+    if isinstance(value, str):
+        return value.strip().lower() in ("true", "1", "yes", "on", "enable", "enabled")
+    return bool(value)
+
+
+def _as_int(value, default: int = 0) -> int:
+    value = _single_value(value)
+    if value is None:
+        return default
+    try:
+        return int(value)
+    except (TypeError, ValueError):
+        return default
+
+
+def _as_float(value, default: float = 0.0) -> float:
+    value = _single_value(value)
+    if value is None:
+        return default
+    try:
+        return float(value)
+    except (TypeError, ValueError):
+        return default
+
+
+def _choice(value, choices, default: str) -> str:
+    choices = tuple(choices or ())
+    value = str(_single_value(value) or "").strip()
+    if value in choices:
+        return value
+    if default in choices:
+        return default
+    return choices[0] if choices else default
+
+
+__all__ = ()
