+++ tests/test_python_backend_analyzer.py
@@ -683,16 +683,14 @@ ignored/
 
         self.assertEqual(analyzer.render_json(report), expected_text)
         self.assertEqual(report["schema_version"], 2)
-        self.assertEqual(report["inventory"]["module_count"], 24)
-        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 24)
-        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 17)
+        self.assertEqual(report["inventory"]["module_count"], 28)
+        self.assertEqual(len(report["registry"]["shipped_python_modules"]), 28)
+        self.assertEqual(len(report["registry"]["runtime_import_closure"]), 23)
         self.assertEqual(report["registry"]["missing_internal_imports"], [])
         self.assertEqual(
             report["registry"]["unreachable_shipped_python_modules"],
             [
-                "easyuse_anima/__init__.py",
                 "easyuse_anima/aio/__init__.py",
-                "easyuse_anima/image/__init__.py",
                 "easyuse_anima/infrastructure/__init__.py",
                 "easyuse_anima/infrastructure/comfy/__init__.py",
                 "easyuse_anima/nodes/__init__.py",
@@ -703,7 +701,11 @@ ignored/
             {
                 "easyuse_anima/__init__.py",
                 "easyuse_anima/aio/__init__.py",
+                "easyuse_anima/common/__init__.py",
+                "easyuse_anima/common/serialization.py",
+                "easyuse_anima/common/values.py",
                 "easyuse_anima/image/__init__.py",
+                "easyuse_anima/image/geometry.py",
                 "easyuse_anima/infrastructure/__init__.py",
                 "easyuse_anima/infrastructure/comfy/__init__.py",
                 "easyuse_anima/nodes/__init__.py",
@@ -712,6 +714,13 @@ ignored/
         )
         self.assertIn("nodes.py", report["registry"]["shipped_python_modules"])
         self.assertIn("api.py", report["registry"]["runtime_import_closure"])
+        self.assertTrue(
+            {
+                "easyuse_anima/common/serialization.py",
+                "easyuse_anima/common/values.py",
+                "easyuse_anima/image/geometry.py",
+            }.issubset(report["registry"]["runtime_import_closure"])
+        )
         self.assertIn(
             "api_contract.py",
             report["registry"]["shipped_python_modules"],
